ISSUE: .CONFIG DISASTER :(
It's seems hard to take care of this issue. See the link provided in
README.md for more details. It seems that there are no clean way to
manage `~/.config`. For example, `ranger` needs to write to
`~/.config/ranger/` in order to launch properly, `emacs` has a config
that's going to be constantly changed directly etc..